refactor: wrap Pierre diff and tree adapters
The Pierre integration should be reusable across the different files-view surfaces instead of being embedded in DiffFile and DiffSidebar. Keeping lifecycle, theme, selection, and demand-context behavior in small Svelte adapters gives callers app-level props and keeps future consumers away from Pierre-specific setup.

Validation: bun run --cwd packages/ui typecheck; bun run --cwd packages/ui lint; bun run --cwd frontend typecheck; bun run --cwd frontend test -- packages/ui/src/components/diff/DiffFile.test.ts packages/ui/src/components/diff/DiffSidebar.test.ts packages/ui/src/components/diff/DiffView.test.ts; bun run --cwd frontend build; browser check on /pulls/github/kenn-io/middleman/383/files
